requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
black==22.1.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.3
django-appconf==1.0.5
django-celery-email==3.0.0
django-ckeditor==6.2.0
django-js-asset==2.0.0
Django-Verify-Email==2.0.1
kombu==5.2.4
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
